Getting neighbours in Agent

+++ src/main/java/Objects/Cell.java
@@ -4,11 +4,11 @@ package Objects;
  * Created by Alexsandr0x.
  */
 public class Cell {
-    public int phero;
+    public double phero;
     public int x;
     public int y;
 
-    public Cell(int x, int y, int phero) {
+    public Cell(int x, int y, double phero) {
         this.phero = phero;
         this.x = x;
         this.y = y;
